MAKE-923: remove mgo backed queues
