+++ queue/remote_ordered_test.go
@@ -13,7 +13,8 @@ import (
 	"github.com/mongodb/grip"
 	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/suite"
-	mgo "gopkg.in/mgo.v2"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 func init() {
@@ -35,22 +36,27 @@ func TestSimpleRemoteOrderedSuiteMongoDB(t *testing.T) {
 
 func (s *SimpleRemoteOrderedSuite) SetupSuite() {
 	name := "test-" + uuid.NewV4().String()
-	uri := "mongodb://localhost"
 	opts := DefaultMongoDBOptions()
 	opts.DB = "amboy_test"
 	s.driverConstructor = func() Driver {
-		return NewMgoDriver(name, opts)
+		return NewMongoDriver(name, opts)
 	}
 
 	s.tearDown = func() error {
-		session, err := mgo.Dial(uri)
-		defer session.Close()
+		ctx, cancel := context.WithCancel(context.Background())
+		defer cancel()
 
+		client, err := mongo.NewClient(options.Client().ApplyURI("mongodb://localhost:27017").SetConnectTimeout(time.Second))
 		if err != nil {
 			return err
 		}
 
-		return session.DB("amboy_test").C(addJobsSuffix(name)).DropCollection()
+		if err := client.Connect(ctx); err != nil {
+			return err
+		}
+		defer client.Disconnect(ctx)
+
+		return client.Database("amboy_test").Collection(addJobsSuffix(name)).Drop(ctx)
 	}
 }
